% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towercount_package.R
\docType{data}
\name{datetime2015}
\alias{datetime2015}
\title{Dataset: Datetime 2015}
\format{numeric}
\usage{
data(datetime2015)
}
\description{
A vector of decimal datetimes from a 2015 count
}
\keyword{datasets}
