% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_few_more_funcs.R
\name{decdatetime}
\alias{decdatetime}
\title{Converting From a Decimal Datetime}
\usage{
decdatetime(datetime, origin = "1899-12-30")
}
\arguments{
\item{datetime}{A vector of decimal datetime}

\item{origin}{A character string specifying the origin of the date numbering system.  This must be structured as "yyyy-mm-dd".  Defaults to \code{"1899-12-30"}}
}
\value{
A two-element list:
\itemize{
\item{\code{$date: }}{Dates in calendar date format, see \link{as.Date}.}
\item{\code{$hour: }}{Hour block recorded as an integer between 0 and 23.}}
}
\description{
Separates the date and hour components of a single column of decimal datetime.
}
\examples{
data(datetime2015)
head(datetime2015)

date_hour <- decdatetime(datetime2015)
head(date_hour$date)
head(date_hour$hour)
}
\author{
Matt Tyers
}
