% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towercount_package.R
\docType{data}
\name{Gulk_2014}
\alias{Gulk_2014}
\title{Dataset: Gulkana Counts 2014}
\format{A data frame with the following fields:
\itemize{
\item{\code{$date: }}{Count dates in calendar date format}
\item{\code{$hour: }}{Hour blocks for each count}
\item{\code{$clarity: }}{Clarity values for each count}
\item{\code{$chin: }}{Chinook salmon counts for each hour block}
\item{\code{$sock: }}{Sockeye salmon counts for each hour block}}}
\usage{
data(Gulk_2014)
}
\description{
A data frame of tower counts from 2014.  There were some substantial high water events this year, compromising counts.
}
\keyword{datasets}
