% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcountmatrix.R
\name{plotcountmatrix}
\alias{plotcountmatrix}
\title{Plot Count Matrix}
\usage{
plotcountmatrix(countsummary, cex = 0.6, cexsummary = 0.7,
  showsummary = TRUE, showinterp = TRUE, ...)
}
\arguments{
\item{countsummary}{A countsummary object from \link{towercountsummary}.}

\item{cex}{The character expansion factor to use for the counts.  Defaults to 0.6.}

\item{cexsummary}{The character expansion factor to use for the summary at bottom right.  Defaults to 0.7.}

\item{showsummary}{Whether to show the summary values for the abundance estimate.  Defaults to TRUE.}

\item{...}{Additional plotting arguments.}
}
\description{
Creates a plot of the count matrix (see \link{towercountsummary}) in which daily counts are color-coded by water clarity.  Any hourly interpolations that differ from the raw counts are highlighted in red.

Also included in the plot are empirical and cumulative run density by date (shown at right) and hour (shown at bottom).  The minimum continuous period containing more than 80 percent of the run is highlighted in green on the hourly density plot at the bottom, and on the hour labels at the top.

At bottom right, the abundance estimate and SE are given, as are an approximate 95 percent Wald confidence interval.
}
\examples{
data(Gulk_2015, Gulk_2014)

chin2015 <- towercountsummary(date=Gulk_2015$date, hour=Gulk_2015$hour,
     count=Gulk_2015$chin, clarity=Gulk_2015$clarity)
plotcountmatrix(chin2015)

sock2015 <- towercountsummary(date=Gulk_2015$date, hour=Gulk_2015$hour,
     count=Gulk_2015$sock, clarity=Gulk_2015$clarity)
plotcountmatrix(sock2015)

chin2014 <- towercountsummary(date=Gulk_2014$date, hour=Gulk_2014$hour,
     count=Gulk_2014$chin, clarity=Gulk_2014$clarity)
plotcountmatrix(chin2014)
}
\author{
Matt Tyers
}
