% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_few_more_funcs.R
\name{savesummary}
\alias{savesummary}
\title{Save Summary}
\usage{
savesummary(countsummary, summaryname = NULL)
}
\arguments{
\item{countsummary}{A countsummary object from \link{towercountsummary}.}

\item{summaryname}{The name to use in place of \code{NAME} in the file names
given above.  Accepting the default (\code{NULL}) results in the name of the
countsummary object being used.}
}
\description{
Saves output from a countsummary object as a set of .csv files,
 in the current working directory.

 \itemize{ \item{\code{NAME_counts_calcs.csv: }}{All counts (including
 interpolations) and calculations leading to expanded estimate and variance
 calculations.}
 \item{\code{NAME_rawcounts.csv: }}{All recorded visual counts.}
 \item{\code{NAME_clarity.csv: }}{A matrix of clarity values
 corresponding to the count matrices.}
 \item{\code{NAME_hourprops.csv:
 }}{The run proportion and cumulative run proportion by hour, as well as the
 interval considered to be the peak daily run.}
 \item{\code{NAME_summary.csv:
 }}{Summary values: the escapement estimate and its associated variance,
 standard error, and 95 percent Wald confidence interval endpoints; also the
 total net number of fish visually counted, the number of periods
 successfully counted, the number of periods missed due to poor visibility,
 the total number of count days, and the number of complete count days with no
 compromised counts.} }
}
\examples{
data(Gulk_2015)

chin2015 <- towercountsummary(date=Gulk_2015$date, hour=Gulk_2015$hour,
     count=Gulk_2015$chin, clarity=Gulk_2015$clarity)
savesummary(chin2015,summaryname="Gulk_chin_2015")
}
\author{
Matt Tyers
}
