% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towercount_package.R
\docType{package}
\name{towercount}
\alias{towercount}
\alias{towercount-package}
\title{Escapement Estimates for River Tower Counts}
\description{
Provides escapement estimates with associated variance for river tower counts.  Also included are functions to visualize and write summary tables.
}
\details{
Version: 0.1

Date: 2015-10-23

License: GPL (>=2)
}
\author{
Matt Tyers

Maintainer: Matt Tyers <matt.tyers@alaska.gov>
}
